refactor: extract network deletion logic
